#3<refact>: Map코드 파일분리
