+++ SixTech/Map/Utilities/ImageAnnotationView.swift
